Bug fix: voxels with variance estimates = 0 are now masked
Issue:
- For some voxels having very (very) small values for the residuals,
the variance estimates of some visit category may be 0 (as it is
obtained from the squared of the residuals computed as 0 by Matlab).
These null variances may cause the failure of the toolbox.
Nevertheless, such a kind of voxels were always observed in the
background far away from the brain (so are of no interest).

Fix:
- The toolbox now automatically masks such a kind of voxels