Set the p-values of null or negative "scaled" F-scores to 1 instead of 0
In older versions of swe_cp_WB, p(scoreTmp == 0) = 1. For some reason, it was changed to p(scoreTmp == 0) = 0. That does not seem correct as small F-scores should have high p-values. Now, I am suggesting to change it to p(scoreTmp <= 0) = 1 with "a lesser than or equal to" comparison.  The latter is used as an additional protection against float comparison equality faillure.