Fix a bug where the post-hoc masking of parametric analyses was not working
It seems that a piece of code (539-559 - old version) was written for WB analysis only. However, it was also visited for parametric analysis and it was automaticaly setting the post-hoc masking to null. Thus, an else statement has been added to ensure that this piece of code is visited only by WB analyses.