[WIP] start handling correctly the post-hoc masking for WB
The main goal of this commit is to make sure that the treshold are always computed based on the images before post-hoc masking. We also start looking at recomputing the clusterwise p-values according to the new cluster sizes, but this is a wip as I need to fix a more urgent issue somewhere else in the code.

A typo was also corrected