save edf in xSwE when the naive edf is used, Vedf otherwise
Before, edf was saved for all parametric analysis and Vedf was saved for all analysis using the modified SwE. That seems weird as the scalar value edf is only relevant for a naive estimation of the numbers of degree of freedom constant across (i.e. when SwE.dof.dof_type == 0). In addition, in all the other cases (i.e. when SwE.dof.dof_type > 0), the numbers of degree of freedom is potentially varying accross voxels and saved into an image Vedf. Thus we should  use a condition on SwE.dof.dof_type to decide what to save.