Deactivate Approx I for the modified SwE and the WB
This could be reactivated later. However, I feel that this approx. should be disabled altogether because Approx III always works as well as or better than Approx I. The reason to deactivate it is that it is was only validated for the WB classic SwE and this would be time consuming to do it as well for the modified SwE